Remove unnecessary EnumChildProc code
The text area child window is automatically resized by vim when
SetWindowPos is called.  There is no need to resize the child window
again afterwards.

However, it would be nice to center the child window within the extra
margin space.  This requires knowledge of the text area's actual working
height and width.  Might not be able to do this from a plugin.